requirements.txt
-e .
aws-cdk-lib
constructs>=10.0.0,<11.0.0
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_s3
aws-cdk.aws_iam
aws-cdk.aws_autoscaling
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_ssm
aws_cdk.aws_secretsmanager
aws_cdk.aws_rds
aws_cdk.aws_sns
aws_cdk.aws_sns_subscriptions
aws_cdk.aws_sqs
aws_cdk.aws_lambda
aws_cdk.aws_logs
aws_cdk.aws_events
aws_cdk.aws_events_targets
aws_cdk.aws_dynamodb
aws_cdk.aws_cloudwatch
aws_cdk.aws_cloudwatch_actions
aws_cdk.aws_apigateway
aws_cdk.aws_s3_deployment
aws_cdk.aws_cloudfront
aws_cdk.aws_s3_notifications
aws_cdk.aws_kinesis
aws_cdk.aws_lambda_event_sources
aws_cdk.aws_ecs
aws_cdk.aws_ecs_patterns